+++ moor/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Support custom, existing classes for rows! See the `@UseRowClass` annotation
   for details.
 - Add `CASE WHEN` expressions with the `caseMatch` method on `Expression`
+- Add `insertReturning` which returns the generated row after an insert
 - On supported platforms, cancel pending stream selects when the stream is disposed
   - `moor_flutter` is supported
   - `moor/ffi` is supported when used on a background isolate